more fixes after review
